pyproject.toml
[tool.poetry]
name = "repl_nix_Info-PSIET"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.6.3"
numpy = "^1.24.1"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
